Restructured the combat system into 'logic' and 'widgets'
