--- repository/scripts/combat/action-bar/bootstrap.js
@@ -1,12 +1,0 @@
-/**
- * Module to initialise the combat abilities system script bindings.
- */
-module.exports = function (scriptManager) {
-	var modules = [
-	    require('./widget')
-	];
-	
-	for (var i in modules) {
-		modules[i].init(scriptManager);
-	}
-};
