+++ repository/scripts/combat/widgets/defence-book.js
@@ -26,8 +26,9 @@ var varbit = require('../../core/var/bit');
 var config = require('../../core/config');
 var widget = require('../../widget');
 
-var logic = require('./logic');
-var actionBar = require('../action-bar/widget');
+var abilities = require('../logic/abilities');
+var actionBar = require('./action-bar');
+var common = require('./common');
 
 /** 
  * @author Im Frizzy <skype:kfriz1998>
@@ -43,44 +44,41 @@ module.exports = (function () {
 	
 	function init (scriptManager) {
 		//Script 8426 = ability book options
-		
-		//Melee ability book
-		scriptManager.bind(EventType.IF_OPEN, 1460, function (ctx) {
-			widget.setEvents(ctx.player, 1460, 1, 0, 187, 10320902);
-			widget.setEvents(ctx.player, 1460, 5, 0, 16, 2);
+		scriptManager.bind(EventType.IF_OPEN, 1449, function (ctx) {
+			widget.setEvents(ctx.player, 1449, 1, 0, 189, 10320902);
+			widget.setEvents(ctx.player, 1449, 7, 0, 16, 2);
 		});
 		
-		scriptManager.bind(EventType.IF_BUTTON1, component(1460, 5), function (ctx) {
+		scriptManager.bind(EventType.IF_BUTTON1, component(1449, 7), function (ctx) {
 			if (ctx.slot === 11) {
 				//Toggle hide
 				varbit(ctx.player, 27344, !varbit(ctx.player, 27344) ? 1 : 0);
 			} else {
-				//Save selected melee tab
-				varbit(ctx.player, 18787, logic.tabIdFromSlot(ctx.slot));
+				//Save selected defence tab
+				varbit(ctx.player, 18793, common.tabIdFromSlot(ctx.slot));
 			}
 		});
 		
-		scriptManager.bind(EventType.IF_BUTTON1, component(1460, 1), function (ctx) {
-			//Script 8437 = ability enums
+		scriptManager.bind(EventType.IF_BUTTON1, component(1449, 1), function (ctx) {
 			var enumId = -1;
-			switch (varbit(ctx.player, 18787)) {
-			case 0://Attack
-				enumId = 6734;
+			switch (varbit(ctx.player, 18793)) {
+			case 0://Defence
+				enumId = 6736;
 				break;
-			case 1://Strength
-				enumId = 6735;
+			case 1://Constitution
+				enumId = 6737;
 				break;
 			default:
-				throw "Unsupported tab: "+varbit(ctx.player, 18787);
+				throw "Unsupported tab: "+varbit(ctx.player, 18793);
 			}
 			var abilityId = config.enumValue(enumId, ctx.slot);
-			logic.runAbility(ctx.player, abilityId);
+			abilities.run(ctx.player, abilityId);
 		});
 		
-		scriptManager.bind(EventType.IF_DRAG, component(1460, 1), function (ctx) {
+		scriptManager.bind(EventType.IF_DRAG, component(1449, 1), function (ctx) {
 			var hash = ctx.toHash;
 			if (widget.getId(hash) == 1430) {
-				actionBar.dragOnto(ctx.player, hash, varbit(ctx.player, 18787) ? 2 : 1, ctx.fromslot);
+				actionBar.dragOnto(ctx.player, hash, varbit(ctx.player, 18793) ? 4 : 3, ctx.fromslot);
 			}
 		});
 	}
