+++ repository/scripts/combat/widgets/ranged-book.js
@@ -21,13 +21,12 @@
  */
 /* globals EventType */
 var component = require('../../widget/component');
-var varbit = require('../../core/var/bit');
 
 var config = require('../../core/config');
 var widget = require('../../widget');
 
-var logic = require('./logic');
-var actionBar = require('../action-bar/widget');
+var abilities = require('../logic/abilities');
+var actionBar = require('./action-bar');
 
 /** 
  * @author Im Frizzy <skype:kfriz1998>
@@ -43,41 +42,20 @@ module.exports = (function () {
 	
 	function init (scriptManager) {
 		//Script 8426 = ability book options
-		scriptManager.bind(EventType.IF_OPEN, 1449, function (ctx) {
-			widget.setEvents(ctx.player, 1449, 1, 0, 189, 10320902);
-			widget.setEvents(ctx.player, 1449, 7, 0, 16, 2);
+		scriptManager.bind(EventType.IF_OPEN, 1452, function (ctx) {
+			widget.setEvents(ctx.player, 1452, 1, 0, 187, 10320902);
+			widget.setEvents(ctx.player, 1452, 7, 0, 16, 2);
 		});
 		
-		scriptManager.bind(EventType.IF_BUTTON1, component(1449, 7), function (ctx) {
-			if (ctx.slot === 11) {
-				//Toggle hide
-				varbit(ctx.player, 27344, !varbit(ctx.player, 27344) ? 1 : 0);
-			} else {
-				//Save selected defence tab
-				varbit(ctx.player, 18793, logic.tabIdFromSlot(ctx.slot));
-			}
+		scriptManager.bind(EventType.IF_BUTTON1, component(1452, 1), function (ctx) {
+			var abilityId = config.enumValue(6738, ctx.slot);
+			abilities.run(ctx.player, abilityId);
 		});
 		
-		scriptManager.bind(EventType.IF_BUTTON1, component(1449, 1), function (ctx) {
-			var enumId = -1;
-			switch (varbit(ctx.player, 18793)) {
-			case 0://Defence
-				enumId = 6736;
-				break;
-			case 1://Constitution
-				enumId = 6737;
-				break;
-			default:
-				throw "Unsupported tab: "+varbit(ctx.player, 18793);
-			}
-			var abilityId = config.enumValue(enumId, ctx.slot);
-			logic.runAbility(ctx.player, abilityId);
-		});
-		
-		scriptManager.bind(EventType.IF_DRAG, component(1449, 1), function (ctx) {
+		scriptManager.bind(EventType.IF_DRAG, component(1452, 1), function (ctx) {
 			var hash = ctx.toHash;
 			if (widget.getId(hash) == 1430) {
-				actionBar.dragOnto(ctx.player, hash, varbit(ctx.player, 18793) ? 4 : 3, ctx.fromslot);
+				actionBar.dragOnto(ctx.player, hash, 5, ctx.fromslot);
 			}
 		});
 	}
