+++ repository/scripts/global-bootstrap.js
@@ -18,8 +18,7 @@ function getAllModules () {// jshint ignore:line
 		'area/kandarin',
 		'chat',
 		'clan',
-		'combat/abilities',
-		'combat/action-bar',
+		'combat/widgets',
 		'inv',
 		'shop',
 		'trade',
